// frontend-admin/src/api/generated/model/licenseType.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export type LicenseType = typeof LicenseType[keyof typeof LicenseType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LicenseType = {
  Trial: 'Trial',
  Starter: 'Starter',
  Business: 'Business',
  Plus: 'Plus',
} as const;