// frontend-admin/src/api/generated/model/tenantStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API â€” RKSV-compliant POS / Admin / shared Auth surfaces
 * OpenAPI spec version: v1
 */

export type TenantStatus = typeof TenantStatus[keyof typeof TenantStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TenantStatus = {
  Lead: 'Lead',
  InOnboarding: 'InOnboarding',
  Active: 'Active',
  Suspended: 'Suspended',
  Cancelled: 'Cancelled',
  Archived: 'Archived',
} as const;